requirements.txt
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
streamlit>=1.28.0
pandas>=1.5.0
transformers>=4.30.0
timm>=0.9.0
open-clip-torch>=2.20.0
pillow>=9.0.0
